+++ adapters/src/main/java/io/realm/RealmBaseAdapter.java
@@ -33,15 +33,21 @@ import android.widget.BaseAdapter;
  * closed. Trying to access Realm objects will at this point also result in a {@code IllegalStateException}.
  */
 public abstract class RealmBaseAdapter<T extends RealmModel> extends BaseAdapter {
-
+    @Nullable
+    @Deprecated
     protected LayoutInflater inflater;
     @Nullable
     protected OrderedRealmCollection<T> adapterData;
-    @NonNull
+    @Nullable
+    @Deprecated
     protected Context context;
     private final RealmChangeListener<? extends BaseRealm> listener;
 
+    @Deprecated
     public RealmBaseAdapter(@NonNull Context context, @Nullable OrderedRealmCollection<T> data) {
+        if (data != null && !data.isManaged())
+            throw new IllegalStateException("Only use this adapter with managed list, " +
+                    "for un-managed lists you can just use the BaseAdapter");
         //noinspection ConstantConditions
         if (context == null) {
             throw new IllegalArgumentException("Context cannot be null");
@@ -49,6 +55,23 @@ public abstract class RealmBaseAdapter<T extends RealmModel> extends BaseAdapter
         this.context = context;
         this.adapterData = data;
         this.inflater = LayoutInflater.from(context);
+        this.listener = new RealmChangeListener<BaseRealm>() {
+            @Override
+            public void onChange(BaseRealm results) {
+                notifyDataSetChanged();
+            }
+        };
+
+        if (data != null) {
+            addListener(data);
+        }
+    }
+
+    public RealmBaseAdapter(@Nullable OrderedRealmCollection<T> data) {
+        if (data != null && !data.isManaged())
+            throw new IllegalStateException("Only use this adapter with managed list, " +
+                    "for un-managed lists you can just use the BaseAdapter");
+        this.adapterData = data;
         this.listener = new RealmChangeListener<BaseRealm>() {
             @Override
             public void onChange(BaseRealm results) {